🐛 fix: fixed param type of yolov5_detector fixed image_transport roi_pub

+++ src/ros2_hik_camera/include/ros2_hik_camera/openvino.h
@@ -33,7 +33,7 @@ public:
   void yolov5_compiled(std::string xml_path, ov::CompiledModel & compiled_model);
   cv::Rect yolov5_detector(
     ov::CompiledModel compiled_model, cv::Mat & input_detect_img, cv::Mat & output_detect_img,
-    vector<cv::Rect> & nms_box, vector<int> & nms_confidence, float size, float conf_thres,
+    vector<cv::Rect> & nms_box, vector<float> & nms_confidence, float size, float conf_thres,
     float nms_thres);
 
 private:
